+++ src/model/comments-model.js
@@ -23,8 +23,8 @@ export default class CommentsModel extends Observable{
     this._notify(updateType, update);
   };
 
-  deleteComment = (updateType, update) => {
-    const index = this.#comments.findIndex((comment) => comment.id === update.id);
+  deleteComment = (updateType, id) => {
+    const index = this.#comments.findIndex((comment) => comment.id === id);
 
     if (index === -1) {
       throw new Error('Can\'t delete unexisting comment');
@@ -34,7 +34,7 @@ export default class CommentsModel extends Observable{
       ...this.#comments.slice(0, index),
       ...this.#comments.slice(index + 1),
     ];
-    this._notify(updateType, update);
+    this._notify(updateType);
   };
 }
